refactor(ext): split monolithic ext.rs into focused submodules
Addresses review item #7 (part 1/3). No behavior change — pure code move.

ext.rs (~637 lines) is now an `ext/` module, one concern per file:
- envelope.rs    — the wire Envelope
- protocol.rs    — the pure core: Event, Effect, Ctx, Inbound, Transition, Protocol
- compute.rs     — the effectful escape hatch (ComputeFn / ComputeServices)
- interpreter.rs — the imperative shell (Interpreter), the only IO boundary
- registry.rs    — type erasure (Handler/Runner) + the namespace router (Extensions)
- mod.rs         — module doc, MaybeSend, and re-exports (public API unchanged)

All external paths (`crate::backend::ext::{Envelope, Effect, Interpreter, ...}`) keep
resolving via the mod.rs re-exports.
